Currently final fix for pushes

+++ src/main/java/org/crama/jelin/repository/GameInitRepository.java
@@ -23,6 +23,6 @@ public interface GameInitRepository {
 
 	boolean removeGameOpponent(Game game, User user);
 	
-	int getExpiredInvites(User user);
+	long getExpiredInvites(User user);
 		
 }
